timeline/events/serializers.py
from rest_framework import serializers
from timeline.events.models import Event
from timeline.users.models import User

class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ('name', )

class EventSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Event
        fields = ('create_date', 'event_type')

